accelerometer: Using different files for different devices
So the code is better organised - just adding code to one .c when adding
new devices is not a good idea.
