+++ src/modules/flow/accelerometer/Makefile
@@ -1,2 +1,3 @@
 obj-$(FLOW_NODE_TYPE_ACCELEROMETER) += accelerometer.mod
-obj-accelerometer-$(FLOW_NODE_TYPE_ACCELEROMETER) := accelerometer.json accelerometer.o
+obj-accelerometer-$(FLOW_NODE_TYPE_ACCELEROMETER) := accelerometer.json \
+	accelerometer.o
